move ToolDetailsPagerAdapter to it's own source file
